AIDS Clinical Trials Group Study 175

ACTG 175 was a randomized clinical trial to compare monotherapy with zidovudine or didanosine with combination therapy with zidovudine and didanosine or zidovudine and zalcitabine in adults infected with the human immunodeficiency virus type I whose CD4 T cell counts were between 200 and 500 per cubic millimeter. Only two of the study arms are included in this practice dataset.

data(ACTG175)

Format

A data frame with 2139 rows and 27 variables:

pidnum

patient's ID number

age

age in years at baseline

wtkg

weight in kg at baseline

hemo

hemophilia (0=no, 1=yes)

msm

Men who have sex with men (0=no, 1=yes)

drugs

history of intravenous drug use (0=no, 1=yes)

karnof

Karnofsky score (on a scale of 0-100)

oprior

non-zidovudine antiretroviral therapy prior to initiation of study treatment (0=no, 1=yes)

z30

zidovudine use in the 30 days prior to treatment initiation (0=no, 1=yes)

zprior

zidovudine use prior to treatment initiation (0=no, 1=yes)

preanti

number of days of previously received antiretroviral therapy

race

race (0=white, 1=non-white)

gender

gender (0=female, 1=male)

str2

antiretroviral history (0=naive, 1=experienced)

strat

antiretroviral history stratification (1='antiretroviral naive', 2='> 1 but ≤q 52 weeks of prior antiretroviral therapy', 3='> 52 weeks')

symptom

symptomatic indicator (0=asymptomatic, 1=symptomatic)

treat

treatment indicator (0=zidovudine only, 1=other therapies)

offtrt

indicator of off-treatment before 96 +/- 5 weeks (0=no,1=yes)

cd40

CD4 T cell count at baseline

cd420

CD4 T cell count at 20 +/- 5 weeks

cd496

CD4 T cell count at 96 +/- 5 weeks (=NA if missing)

r

missing CD4 T cell count at 96 +/- 5 weeks (0=missing, 1=observed)

cd80

CD8 T cell count at baseline

cd820

CD8 T cell count at 20 weeks

cens

indicator of observing the event in days

days

number of days until the first occurrence of: (i) a decline in CD4 T cell count of at least 50 (ii) an event indicating progression to AIDS, or (iii) death.

arms

treatment arm (0=zidovudine, 1=zidovudine and didanosine, 2=zidovudine and zalcitabine, 3=didanosine)

Source

https://cran.r-project.org/web/packages/speff2trial/index.html

Details

The variable days contains right-censored time-to-event observations. The data set includes the following post-randomization covariates: CD4 and CD8 T cell count at 20+/-5 weeks and the indicator of whether or not the patient was taken off-treatment before 96+/-5 weeks.

References

Hammer SM, et al. (1996), "A trial comparing nucleoside monotherapy with combination therapy in HIV-infected adults with CD4 cell counts from 200 to 500 per cubic millimeter.", New England Journal of Medicine, 335: 1081 – 1090.