mdis_chrt: Matrix algebra in base R for covariate balance using Mahalanobis distance of a point to its distribution

This function will compute M-distance for a covariate matrix to its mean

mdis_chrt(x_i)

Arguments

x_i
A vector or covariate matrix

Value

a numeric vector of length equal to x_i

Examples

require(jumble)
df_x <- ACTG175[, c('age', 'gender', 'race', 'msm')]

mdis_chrt(df_x)

#>    [1] 1.1300570 1.3883637 1.0505993 0.9890806 0.8542386 0.8847327 0.9324092
#>    [8] 2.6425066 1.4826899 1.1246045 0.8874008 1.2913748 0.9288526 2.5636425
#>   [15] 0.8558975 1.2975844 1.5799279 0.8847327 2.5352920 1.6795719 0.8847327
#>   [22] 0.9288526 1.1300570 1.1246045 1.6725191 2.7567622 0.8394212 1.5730295
#>   [29] 1.7812182 0.9324092 2.6783716 1.2052795 2.3664663 2.7724729 2.5305880
#>   [36] 0.8542386 3.1598275 2.5307247 2.8451470 2.6566685 2.6395515 1.3818778
#>   [43] 2.5880072 2.7324499 2.4190659 2.7251011 2.3372571 0.9288526 1.2913748
#>   [50] 3.9844123 1.0555383 2.5610138 1.1246045 0.9324092 2.3883424 1.2052795
#>   [57] 2.5031320 1.4826899 2.1132152 0.8874008 0.9890806 1.2052795 1.0555383
#>   [64] 1.3818778 1.2111497 0.8394212 1.2913748 1.4759778 1.3818778 1.6795719
#>   [71] 2.6113205 2.7187304 1.5730295 2.7633861 1.2052795 0.8388577 1.7740365
#>   [78] 0.9890806 1.2913748 1.2913748 2.3316875 1.0555383 1.1300570 2.6735692
#>   [85] 3.4187466 0.8542386 2.2199064 0.8847327 2.1059831 2.6122520 1.2975844
#>   [92] 2.0981689 0.8394212 2.6018200 2.6395515 2.5905724 1.2111497 2.1059831
#>   [99] 3.3725555 1.3818778 2.1201424 0.9324092 1.1246045 1.1246045 3.8529557
#>  [106] 2.3282536 3.9027270 2.1844742 0.9847688 0.9288526 0.8558975 2.2370136
#>  [113] 2.5836452 0.8874008 0.8394212 2.4190659 2.4147289 0.9890806 0.8558975
#>  [120] 0.9324092 1.7812182 1.2975844 1.6795719 1.3883637 2.3892681 0.8394212
#>  [127] 2.4418892 0.8558975 2.7482983 1.2111497 2.1309093 1.2111497 3.3703196
#>  [134] 3.7575917 1.2052795 2.1309093 2.5849826 2.6317229 0.8388577 1.5799279
#>  [141] 1.3883637 1.0505993 0.8558975 2.9378613 2.1132152 1.0505993 2.1309093
#>  [148] 2.3828509 3.0320012 2.3282536 1.2913748 2.8121339 1.0505993 1.0555383
#>  [155] 0.9324092 0.9890806 2.3828509 2.2199064 2.4163243 3.7157493 0.8558975
#>  [162] 1.2111497 0.9890806 3.5314865 3.3307232 0.8394212 0.8874008 0.8558975
#>  [169] 1.2052795 0.8394212 0.8874008 1.4826899 0.9324092 1.5799279 2.5280053
#>  [176] 2.1547291 1.5799279 1.0555383 0.9847688 2.5845020 0.8847327 0.8847327
#>  [183] 1.0505993 2.2607584 0.9288526 1.2913748 0.8388577 1.2913748 1.2913748
#>  [190] 2.1405209 2.0967708 2.1132152 2.3067422 1.2975844 2.3491935 1.2052795
#>  [197] 1.3883637 2.3897605 1.4759778 2.1018015 2.1201424 2.3067422 0.9288526
#>  [204] 1.4759778 0.9288526 1.3818778 1.2913748 2.4382879 0.8388577 0.8542386
#>  [211] 0.9288526 0.9847688 1.3818778 3.0942093 0.8558975 1.1300570 0.8874008
#>  [218] 0.8558975 0.9890806 1.3883637 1.2052795 1.6795719 1.1246045 2.1309093
#>  [225] 2.5364303 2.5352920 1.4759778 0.8847327 1.1246045 0.8874008 2.1018015
#>  [232] 1.1300570 1.1246045 3.2170885 0.8394212 2.8591590 2.3657394 2.3575575
#>  [239] 0.9288526 0.8558975 2.1201424 1.4759778 0.8847327 0.8558975 1.0555383
#>  [246] 0.8874008 2.4163243 2.3883424 0.8388577 0.8394212 0.9847688 0.8394212
#>  [253] 2.5280053 1.1300570 2.5816303 1.3818778 3.7575917 0.8542386 2.1844742
#>  [260] 0.8874008 2.5838855 1.4826899 0.9890806 0.8558975 0.9847688 1.8845429
#>  [267] 1.0555383 1.4759778 1.0555383 2.1547291 0.8394212 0.9890806 2.6030788
#>  [274] 0.8542386 0.8847327 0.9847688 2.1018015 2.5031320 1.1246045 2.1991914
#>  [281] 2.3878799 2.5455363 2.5323210 2.4953428 0.9847688 0.8542386 2.4632922
#>  [288] 1.6725191 1.2111497 1.4759778 0.9847688 1.3818778 2.6181780 2.2199064
#>  [295] 2.3828509 1.5730295 1.2913748 1.1300570 2.1132152 2.5636425 2.6201983
#>  [302] 2.3878799 2.7377106 2.2199064 1.4759778 1.3883637 2.5610138 2.3883424
#>  [309] 0.8394212 1.5730295 0.8394212 1.0505993 3.8016518 2.3316875 0.9847688
#>  [316] 1.3818778 2.6968399 0.8558975 3.0942093 1.2913748 1.1246045 2.5740145
#>  [323] 2.5610138 2.6968399 1.3883637 2.4495007 2.9728391 2.6181780 2.8647650
#>  [330] 2.3067422 2.1309093 0.9288526 2.7800590 1.4759778 1.2111497 1.0555383
#>  [337] 0.8847327 2.4128992 2.3878799 2.5636425 2.4128992 1.8845429 0.9288526
#>  [344] 0.9324092 1.6795719 2.7472985 2.4174425 0.9324092 1.7812182 0.8847327
#>  [351] 0.8874008 0.8542386 0.9890806 3.1940025 1.1300570 1.1300570 0.8847327
#>  [358] 0.8388577 0.8558975 1.1246045 0.8847327 1.2052795 0.8542386 1.0505993
#>  [365] 1.1300570 1.5799279 1.2052795 1.1246045 2.0952304 2.3372571 1.3818778
#>  [372] 0.9288526 1.5730295 1.3883637 2.1059831 1.7812182 1.2111497 1.1246045
#>  [379] 1.7812182 2.7472985 2.4418892 1.5799279 3.4187466 1.2111497 2.7567622
#>  [386] 1.2052795 0.8394212 0.8394212 2.5798457 1.3883637 1.0555383 0.8558975
#>  [393] 1.2975844 0.9847688 1.0555383 2.2370136 2.7482983 2.0981689 0.9324092
#>  [400] 2.4163243 1.2975844 0.9890806 2.3375724 2.3315821 2.8576888 1.3818778
#>  [407] 1.0555383 2.1059831 1.3818778 3.6649937 0.8394212 1.0505993 2.1991914
#>  [414] 1.3818778 1.0505993 2.4128992 1.4759778 2.3828509 1.5799279 2.5880072
#>  [421] 0.9288526 1.4759778 2.3664663 2.4495007 2.5352920 1.2052795 2.5905724
#>  [428] 0.8388577 0.8388577 2.6658198 0.9847688 0.9288526 2.5474256 2.4174425
#>  [435] 0.8542386 0.8874008 1.6725191 2.4016503 1.0505993 1.5799279 1.6725191
#>  [442] 1.2052795 2.5474256 1.0505993 1.3883637 1.1246045 2.4891449 2.3810739
#>  [449] 2.3892681 1.7812182 0.8394212 0.8847327 2.1844742 1.1300570 2.1669432
#>  [456] 1.1246045 2.5411756 2.3985307 1.2111497 2.4363638 1.1300570 1.3818778
#>  [463] 2.6018200 2.1018015 1.3818778 2.2199064 1.4759778 1.2052795 0.9324092
#>  [470] 1.1246045 0.8394212 2.5816303 3.8218865 1.1246045 3.0496379 2.8745487
#>  [477] 0.9890806 2.3100707 2.3985307 2.6317229 2.4363638 1.2975844 2.5849826
#>  [484] 2.6395515 3.1050408 2.5740145 3.7722547 2.6018200 0.8388577 0.8542386
#>  [491] 2.6122520 1.1246045 3.0943944 0.8874008 3.8113805 2.4418892 3.6496182
#>  [498] 2.4363638 2.6310345 0.9847688 1.3818778 0.9288526 0.9847688 1.0505993
#>  [505] 2.1309093 0.8388577 1.5730295 1.2111497 0.9324092 2.6395515 2.2199064
#>  [512] 1.4826899 0.8388577 0.8874008 2.3897605 2.5359522 2.3828509 2.5880072
#>  [519] 2.4190659 2.3834792 2.5838855 2.5636425 2.6566685 2.4190659 2.5474256
#>  [526] 2.3985307 2.3834792 2.3315821 2.7377106 2.6201983 2.6783716 2.5307247
#>  [533] 2.3985307 2.5303444 2.5411756 2.5031320 1.0555383 2.7724729 2.7251011
#>  [540] 2.4016503 2.6968399 0.8542386 1.1246045 1.0505993 0.8542386 1.2111497
#>  [547] 0.8394212 1.4759778 0.8874008 3.0820520 2.1309093 1.1246045 2.5455363
#>  [554] 1.6795719 0.9288526 0.9847688 1.6795719 2.3100707 0.8394212 1.0505993
#>  [561] 2.0877700 2.4128992 2.3375724 2.1547291 2.2607584 1.1246045 2.1309093
#>  [568] 2.6566685 0.8394212 1.1246045 2.1309093 0.9288526 0.8394212 1.1246045
#>  [575] 1.7740365 3.8709204 2.9102722 1.2913748 1.2052795 2.5905724 2.3067422
#>  [582] 2.3657394 1.0555383 3.0203742 1.2111497 2.5636425 0.9847688 3.0346887
#>  [589] 2.7324499 2.5816303 2.2607584 2.5411756 2.6319917 1.0555383 2.5455363
#>  [596] 2.7377106 1.3818778 0.8394212 1.0555383 0.8847327 1.2052795 0.9324092
#>  [603] 0.8847327 0.9288526 0.8847327 1.2975844 1.0555383 2.5636425 2.5303444
#>  [610] 0.8542386 2.2022069 1.1246045 2.1405209 1.4759778 1.0555383 0.8558975
#>  [617] 2.1018015 1.5799279 1.4759778 1.2052795 2.7450618 2.7471644 2.7588640
#>  [624] 2.6425066 1.2052795 3.2487641 0.8388577 2.4418892 0.8542386 2.6113205
#>  [631] 2.5323210 2.5836452 2.5975618 2.6201983 0.9890806 2.5303444 0.9890806
#>  [638] 2.8049391 2.4995940 2.5849826 1.0555383 0.8542386 2.8121339 0.8388577
#>  [645] 0.9847688 0.8847327 1.1300570 0.9847688 0.8874008 1.8845429 3.6444413
#>  [652] 2.4953428 1.0505993 2.8570292 2.6783716 0.8874008 0.8542386 1.0555383
#>  [659] 1.2975844 1.6795719 0.9288526 2.5323210 2.4699702 1.2052795 0.8847327
#>  [666] 0.8388577 0.8558975 2.2022069 0.8542386 1.2913748 0.8847327 2.7251011
#>  [673] 1.2111497 0.8542386 2.5031320 0.9288526 0.8394212 2.0967708 2.6968399
#>  [680] 1.8845429 1.9892845 2.1547291 1.2975844 0.9890806 0.9288526 0.8394212
#>  [687] 1.0555383 0.8847327 1.3818778 1.1300570 2.1844742 2.5646986 2.4632922
#>  [694] 1.2052795 1.3818778 1.4826899 1.0505993 0.9847688 1.9819024 1.0505993
#>  [701] 2.1059831 2.4724632 0.8847327 1.2111497 2.0952304 0.9288526 0.8542386
#>  [708] 1.4759778 0.8394212 0.8847327 1.1300570 1.3818778 1.4759778 1.0505993
#>  [715] 0.8394212 2.7833478 1.2052795 1.2975844 1.2052795 2.1405209 1.1300570
#>  [722] 0.8388577 0.8874008 1.1300570 0.8874008 1.0505993 2.3315821 1.3818778
#>  [729] 1.1246045 0.9847688 2.3282536 3.7954147 2.2607584 2.7588640 2.5455363
#>  [736] 2.3375724 0.8388577 1.7740365 2.4128992 2.0877700 2.6018200 2.5816303
#>  [743] 1.5730295 0.9288526 2.2607584 0.8388577 0.9324092 0.8874008 1.1246045
#>  [750] 0.9324092 0.9890806 0.8558975 0.8874008 1.2052795 0.9288526 1.4826899
#>  [757] 1.1300570 2.5845020 1.3818778 0.8394212 2.4724632 1.2913748 2.5610138
#>  [764] 1.1246045 2.1132152 1.2975844 2.1669432 2.5031320 0.9847688 2.1844742
#>  [771] 2.5880072 0.8542386 2.0967708 0.8558975 1.6795719 0.8542386 0.8847327
#>  [778] 2.5610138 0.9324092 2.1132152 1.0505993 2.4724632 0.9288526 0.9847688
#>  [785] 2.6706393 2.2199064 0.8394212 1.2111497 2.6855189 2.3664663 2.4190659
#>  [792] 2.5280053 2.3491935 0.8874008 1.3883637 1.2052795 0.8394212 1.2913748
#>  [799] 2.3878799 2.9940502 2.3897605 2.5323210 2.4876638 1.1246045 3.0914288
#>  [806] 2.3315821 2.4876638 1.4826899 2.8649983 2.4495007 2.4953428 2.5474256
#>  [813] 2.7633861 2.5455363 3.8891267 0.9890806 2.3315821 0.8388577 2.7633861
#>  [820] 1.2052795 2.5836452 2.8995959 2.6201983 3.1099637 0.8558975 0.9324092
#>  [827] 2.6030788 1.3818778 2.3315821 2.4163243 2.5322380 2.6377686 3.1718115
#>  [834] 2.3883424 0.8847327 2.3878799 2.7251011 2.1844742 2.5838855 2.8593223
#>  [841] 1.5730295 3.7004453 1.5730295 0.8394212 1.3818778 0.8394212 1.5730295
#>  [848] 2.7633861 2.3067422 2.5646986 2.3834792 2.3985307 4.0471271 0.9288526
#>  [855] 2.6783716 1.5730295 2.8570292 1.3883637 1.4759778 2.3372571 1.8772528
#>  [862] 0.9847688 1.7740365 1.6725191 2.3878799 0.9288526 1.2052795 0.8558975
#>  [869] 2.4508043 2.5836452 1.3818778 2.6425066 1.5730295 2.4699702 2.3810739
#>  [876] 1.8772528 2.0967708 0.9847688 2.9728391 1.2111497 2.5838855 2.4876638
#>  [883] 2.3657394 2.3315821 1.0505993 2.6377686 2.7324499 2.5305880 2.9728391
#>  [890] 0.9890806 2.8570292 0.8847327 1.0505993 2.4953428 1.1246045 2.4508043
#>  [897] 2.5031320 1.1300570 0.9890806 2.0952304 0.9890806 1.6795719 0.9324092
#>  [904] 2.2199064 0.8542386 2.4632922 2.9171047 2.5610138 2.4147289 2.4363638
#>  [911] 2.9027148 2.4632922 2.8236216 2.1309093 2.5905724 1.9892845 2.1547291
#>  [918] 0.9890806 2.1991914 2.4632922 2.3878799 2.4363638 2.4190659 2.6181780
#>  [925] 2.3878799 2.4632922 2.5905724 2.8649983 2.3575575 2.4632922 2.5474256
#>  [932] 1.2975844 2.4174425 0.8558975 0.8558975 2.1018015 0.8388577 1.1300570
#>  [939] 1.2052795 2.5307247 0.8874008 0.9324092 0.9288526 1.2111497 2.1201424
#>  [946] 1.1246045 1.5730295 1.2913748 0.9324092 0.9890806 0.8874008 0.9288526
#>  [953] 0.9847688 1.2913748 0.9847688 1.0505993 2.1405209 2.3897605 2.4632922
#>  [960] 0.8874008 2.5610138 2.0967708 1.1246045 2.4699702 2.9728391 2.6968399
#>  [967] 2.0967708 1.0555383 1.2975844 2.1844742 2.2607584 1.2052795 2.4163243
#>  [974] 1.8845429 3.1003574 2.5364303 2.1201424 2.4508043 2.7724729 2.6122520
#>  [981] 2.7187304 2.6566685 2.5636425 2.9721107 1.0505993 2.3575575 3.2170885
#>  [988] 2.5364303 3.0820520 2.5303444 2.6377686 2.8049391 2.6310345 2.9171047
#>  [995] 0.8874008 2.3664663 2.4163243 2.8647650 2.3491935 2.3486705 2.5798457
#> [1002] 2.7472985 2.3316875 2.3315821 2.8105109 2.3892681 3.4146673 3.2054027
#> [1009] 2.8017469 3.0735754 2.3892681 2.3883424 2.3316875 2.6876223 2.7588640
#> [1016] 2.3375724 2.6876223 2.3316875 2.7472985 2.6018200 2.6876223 2.4876638
#> [1023] 2.4174425 3.3337463 2.3892681 3.3337463 2.3491935 2.8995959 3.0190827
#> [1030] 2.8995959 2.4876638 3.2517123 0.9847688 2.8770264 2.5798457 2.8487239
#> [1037] 2.3491935 2.7472985 3.1718115 2.9440353 2.3883424 2.8770264 2.4163243
#> [1044] 2.6317229 2.6317229 2.5845020 2.5305880 2.3657394 2.5322380 2.6876223
#> [1051] 3.0190827 2.9808392 3.1380317 3.1718115