rnd_str: Random assignment by strata

Take a dataset and assign treatment by a specified stratum

rnd_str(df, strata, id)

Arguments

df
a dataframe, containing strata variable

strata
a column in df naming the stratification variable, can be character or numeric

id
A variable identify unit of randomization

Value

dataframe with new column assigning treatment

Examples

df <- jumble::ACTG175
rnd_str(df, symptom, pidnum)

#>      symptom pidnum group
#> 1          0  10124     a
#> 2          0  10140     a
#> 3          0  10165     a
#> 4          0  10190     b
#> 5          0  10229     b
#> 6          0  10361     b
#> 7          0  10378     b
#> 8          0  10389     b
#> 9          0  10668     b
#> 10         0  10716     b
#> 11         0  10721     a
#> 12         0  10896     b
#> 13         0  10900     b
#> 14         0  10907     a
#> 15         0  10914     a
#> 16         0  10915     b
#> 17         0  10917     b
#> 18         0  10923     b
#> 19         0  10935     b
#> 20         0  10938     b
#> 21         0  10950     a
#> 22         0  10958     b
#> 23         0  10962     b
#> 24         0  11059     a
#> 25         0  11128     a
#> 26         0  11157     b
#> 27         0  11164     a
#> 28         0  11167     a
#> 29         0  11171     a
#> 30         0  11179     b
#> 31         0  11180     b
#> 32         0  11187     b
#> 33         0  11201     b
#> 34         0  11358     b
#> 35         0  11369     b
#> 36         0  11435     a
#> 37         0  11438     a
#> 38         0  11446     b
#> 39         0  11449     a
#> 40         0  11454     b
#> 41         0  11468     b
#> 42         0  11470     a
#> 43         0  11475     b
#> 44         0  11479     a
#> 45         0  11499     b
#> 46         0  11515     a
#> 47         0  11651     b
#> 48         0  11652     b
#> 49         0  11654     b
#> 50         0  11656     a
#> 51         0  11661     b
#> 52         0  11665     a
#> 53         0  11668     a
#> 54         0  11669     a
#> 55         0  11672     b
#> 56         0  11676     a
#> 57         0  11959     a
#> 58         0  11961     b
#> 59         0  11971     b
#> 60         0  11973     b
#> 61         0  11974     b
#> 62         0  11975     b
#> 63         0  11976     b
#> 64         0  11981     b
#> 65         0  11983     b
#> 66         0  11984     a
#> 67         0  11986     a
#> 68         0  11987     b
#> 69         0  11997     b
#> 70         0  12003     a
#> 71         0  12008     a
#> 72         0  20103     b
#> 73         0  20187     a
#> 74         0  20386     b
#> 75         0  20438     b
#> 76         0  20505     a
#> 77         0  20509     a
#> 78         0  20678     b
#> 79         0  20682     b
#> 80         0  20683     a
#> 81         0  20689     b
#> 82         0  20696     a
#> 83         0  20702     b
#> 84         0  20705     a
#> 85         0  30134     a
#> 86         0  30137     a
#> 87         0  30349     b
#> 88         0  30350     a
#> 89         0  30353     b
#> 90         0  40409     a
#> 91         0  41271     a
#> 92         0  41650     b
#> 93         0  41655     b
#> 94         0  41658     a
#> 95         0  41673     b
#> 96         0  50137     b
#> 97         0  50185     a
#> 98         0  50200     b
#> 99         0  50225     a
#> 100        0  50260     b
#> 101        0  50321     a
#> 102        0  50344     a
#> 103        0  50474     a
#> 104        0  50566     b
#> 105        0  50571     b
#> 106        0  50577     a
#> 107        0  50578     b
#> 108        0  50580     b
#> 109        0  50582     a
#> 110        0  50583     b
#> 111        0  50587     a
#> 112        0  50603     b
#> 113        0  50607     a
#> 114        0  50608     b
#> 115        0  50612     b
#> 116        0  50615     b
#> 117        0  50620     b
#> 118        0  50622     b
#> 119        0  50624     a
#> 120        0  50629     a
#> 121        0  50633     b
#> 122        0  50634     b
#> 123        0  50638     b
#> 124        0  50639     b
#> 125        0  50641     a
#> 126        0  50643     b
#> 127        0  50651     b
#> 128        0  50652     b
#> 129        0  50654     b
#> 130        0  50655     a
#> 131        0  50660     b
#> 132        0  50662     a
#> 133        0  50663     a
#> 134        0  60245     b
#> 135        0  60260     b
#> 136        0  60267     a
#> 137        0  60295     b
#> 138        0  60335     a
#> 139        0  60451     b
#> 140        0  60613     a
#> 141        0  60615     b
#> 142        0  60617     a
#> 143        0  60618     b
#> 144        0  60649     b
#> 145        0  60888     b
#> 146        0  60889     b
#> 147        0  60892     a
#> 148        0  60896     b
#> 149        0  60900     a
#> 150        0  60902     a
#> 151        0  60921     b
#> 152        0  60959     b
#> 153        0  60960     a
#> 154        0  60966     b
#> 155        0  60967     a
#> 156        0  60972     a
#> 157        0  60977     a
#> 158        0  70255     a
#> 159        0  70264     a
#> 160        0  70308     b
#> 161        0  70311     b
#> 162        0  70330     b
#> 163        0  70455     a
#> 164        0  70463     a
#> 165        0  70466     b
#> 166        0  70479     a
#> 167        0  70521     a
#> 168        0  70528     b
#> 169        0  70643     b
#> 170        0  70794     b
#> 171        0  71328     a
#> 172        0  71329     b
#> 173        0  71331     b
#> 174        0  71333     b
#> 175        0  71341     b
#> 176        0  71342     a
#> 177        0  71343     a
#> 178        0  71347     b
#> 179        0  71353     b
#> 180        0  71361     a
#> 181        0  71366     a
#> 182        0  71369     a
#> 183        0  71370     b
#> 184        0  71371     a
#> 185        0  71376     b
#> 186        0  71379     b
#> 187        0  71380     b
#> 188        0  71386     a
#> 189        0  71389     b
#> 190        0  80272     b
#> 191        0  80392     b
#> 192        0  80466     a
#> 193        0  80748     b
#> 194        0  80933     a
#> 195        0  81021     a
#> 196        0  81103     a
#> 197        0  81105     b
#> 198        0  81127     a
#> 199        0  81134     b
#> 200        0  81135     a
#> 201        0  81136     a
#> 202        0  81138     b
#> 203        0  81140     b
#> 204        0  81141     a
#> 205        0  81143     a
#> 206        0  81144     b
#> 207        0  81145     b
#> 208        0  81445     b
#> 209        0  81446     b
#> 210        0  81451     b
#> 211        0  81456     a
#> 212        0  81458     b
#> 213        0  81459     a
#> 214        0  81460     b
#> 215        0  81461     b
#> 216        0  81463     a
#> 217        0  90278     b
#> 218        0  90292     a
#> 219        0  90297     a
#> 220        0  90300     b
#> 221        0  90335     b
#> 222        0  90441     b
#> 223        0  90668     a
#> 224        0  90669     a
#> 225        0  90672     a
#> 226        0  90674     b
#> 227        0  90675     a
#> 228        0  90677     b
#> 229        0  90681     b
#> 230        0  90682     a
#> 231        0  90691     a
#> 232        0  90693     b
#> 233        0  90698     a
#> 234        0  90699     a
#> 235        0  90700     a
#> 236        0  90703     b
#> 237        0  90707     b
#> 238        0  90712     b
#> 239        0 110154     a
#> 240        0 110384     a
#> 241        0 110581     b
#> 242        0 110648     a
#> 243        0 110683     b
#> 244        0 110697     b
#> 245        0 110698     b
#> 246        0 110705     a
#> 247        0 110708     b
#> 248        0 110709     a
#> 249        0 110713     b
#> 250        0 110714     b
#> 251        0 110720     a
#> 252        0 110724     a
#> 253        0 110730     a
#> 254        0 110733     b
#> 255        0 110741     a
#> 256        0 110743     a
#> 257        0 110753     b
#> 258        0 110754     a
#> 259        0 110762     a
#> 260        0 110764     a
#> 261        0 120142     a
#> 262        0 120328     b
#> 263        0 120368     a
#> 264        0 120892     a
#> 265        0 120896     b
#> 266        0 120897     b
#> 267        0 120898     b
#> 268        0 120905     b
#> 269        0 120910     b
#> 270        0 120911     b
#> 271        0 120918     a
#> 272        0 120927     a
#> 273        0 120936     b
#> 274        0 120940     a
#> 275        0 120948     b
#> 276        0 140187     a
#> 277        0 140232     b
#> 278        0 140400     a
#> 279        0 140714     b
#> 280        0 140718     b
#> 281        0 140731     a
#> 282        0 140741     a
#> 283        0 140742     a
#> 284        0 140744     b
#> 285        0 140745     a
#> 286        0 140750     b
#> 287        0 140753     b
#> 288        0 140754     a
#> 289        0 140756     a
#> 290        0 140758     b
#> 291        0 140761     a
#> 292        0 140764     a
#> 293        0 140765     b
#> 294        0 140773     a
#> 295        0 140775     b
#> 296        0 140778     b
#> 297        0 140780     a
#> 298        0 140783     b
#> 299        0 140787     b
#> 300        0 140789     a
#> 301        0 140791     b
#> 302        0 140792     a
#> 303        0 140795     a
#> 304        0 140797     b
#> 305        0 140802     a
#> 306        0 140809     a
#> 307        0 140811     a
#> 308        0 150073     b
#> 309        0 150089     a
#> 310        0 150144     b
#> 311        0 150147     b
#> 312        0 150159     b
#> 313        0 150163     a
#> 314        0 150166     a
#> 315        0 150214     b
#> 316        0 150225     a
#> 317        0 150227     b
#> 318        0 150236     b
#> 319        0 150272     a
#> 320        0 150422     b
#> 321        0 150425     a
#> 322        0 150427     a
#> 323        0 150429     b
#> 324        0 150432     b
#> 325        0 150438     a
#> 326        0 150439     b
#> 327        0 150444     a
#> 328        0 150447     a
#> 329        0 150450     a
#> 330        0 150452     a
#> 331        0 150454     b
#> 332        0 150456     b
#> 333        0 150458     a
#> 334        0 150459     a
#> 335        0 150465     a
#> 336        0 150473     b
#> 337        0 150474     b
#> 338        0 150477     b
#> 339        0 150482     b
#> 340        0 150484     a
#> 341        0 150490     a
#> 342        0 150495     a
#> 343        0 150497     b
#> 344        0 150502     a
#> 345        0 150503     a
#> 346        0 150507     b
#> 347        0 150510     b
#> 348        0 160111     a
#> 349        0 160142     b
#> 350        0 160204     b
#> 351        0 160223     b
#> 352        0 160344     a
#> 353        0 160375     b
#> 354        0 160376     a
#> 355        0 160378     a
#> 356        0 160381     a
#> 357        0 160384     b
#> 358        0 170222     a
#> 359        0 170251     a
#> 360        0 170356     a
#> 361        0 170398     b
#> 362        0 170635     b
#> 363        0 170640     b
#> 364        0 170645     b
#> 365        0 170646     a
#> 366        0 170648     b
#> 367        0 170651     a
#> 368        0 170653     b
#> 369        0 170662     a
#> 370        0 170686     b
#> 371        0 170846     a
#> 372        0 170868     b
#> 373        0 170888     b
#> 374        0 170896     b
#> 375        0 170945     a
#> 376        0 170949     b
#> 377        0 170963     a
#> 378        0 170972     b
#> 379        0 170987     b
#> 380        0 171001     b
#> 381        0 171040     b
#> 382        0 180091     a
#> 383        0 180156     b
#> 384        0 180220     a
#> 385        0 180315     a
#> 386        0 180394     b
#> 387        0 180686     a
#> 388        0 180701     a
#> 389        0 180704     b
#> 390        0 180712     a
#> 391        0 180876     a
#> 392        0 180917     b
#> 393        0 180919     b
#> 394        0 181173     a
#> 395        0 181285     a
#> 396        0 181304     b
#> 397        0 181309     a
#> 398        0 181317     a
#> 399        0 181318     a
#> 400        0 181319     b
#> 401        0 181322     a
#> 402        0 181325     b
#> 403        0 181352     a
#> 404        0 181960     a
#> 405        0 181968     a
#> 406        0 181971     a
#> 407        0 181982     b
#> 408        0 181994     a
#> 409        0 181996     a
#> 410        0 181997     a
#> 411        0 181999     b
#> 412        0 182010     a
#> 413        0 182017     b
#> 414        0 182020     a
#> 415        0 182028     a
#> 416        0 182030     b
#> 417        0 182037     b
#> 418        0 190192     b
#> 419        0 190193     a
#> 420        0 190195     a
#> 421        0 190204     a
#> 422        0 190210     a
#> 423        0 190277     b
#> 424        0 190281     b
#> 425        0 190386     a
#> 426        0 190394     a
#> 427        0 190395     a
#> 428        0 190399     b
#> 429        0 190403     a
#> 430        0 190405     b
#> 431        0 190407     b
#> 432        0 190419     b
#> 433        0 190526     a
#> 434        0 190529     a
#> 435        0 190566     a
#> 436        0 190597     b
#> 437        0 191604     a
#> 438        0 191605     b
#> 439        0 191620     a
#> 440        0 191629     b
#> 441        0 200080     b
#> 442        0 210039     a
#> 443        0 210141     b
#> 444        0 210293     b
#> 445        0 210331     a
#> 446        0 210404     a
#> 447        0 210407     a
#> 448        0 210561     a
#> 449        0 210601     b
#> 450        0 210623     b
#> 451        0 210766     b
#> 452        0 210800     a
#> 453        0 210865     a
#> 454        0 210883     b
#> 455        0 210919     b
#> 456        0 210970     a
#> 457        0 211007     a
#> 458        0 211033     a
#> 459        0 211040     a
#> 460        0 211057     b
#> 461        0 211065     a
#> 462        0 211074     a
#> 463        0 211086     a
#> 464        0 211235     b
#> 465        0 211253     a
#> 466        0 211255     a
#> 467        0 211258     a
#> 468        0 211272     b
#> 469        0 211281     b
#> 470        0 211283     b
#> 471        0 211284     b
#> 472        0 211314     a
#> 473        0 211333     a
#> 474        0 211344     a
#> 475        0 211370     a
#> 476        0 211377     a
#> 477        0 220081     b
#> 478        0 220120     b
#> 479        0 220130     a
#> 480        0 220423     a
#> 481        0 220424     b
#> 482        0 220425     b
#> 483        0 220431     b
#> 484        0 220437     a
#> 485        0 220438     a
#> 486        0 220451     a
#> 487        0 220458     b
#> 488        0 220478     a
#> 489        0 220489     b
#> 490        0 220491     b
#> 491        0 220492     b
#> 492        0 230082     a
#> 493        0 230109     b
#> 494        0 230233     b
#> 495        0 230238     a
#> 496        0 230546     b
#> 497        0 230568     b
#> 498        0 230606     a
#> 499        0 230753     a
#> 500        0 230822     a
#> 501        0 230832     b
#> 502        0 230844     a
#> 503        0 231018     a
#> 504        0 231021     b
#> 505        0 231030     b
#> 506        0 231031     a
#> 507        0 231052     b
#> 508        0 231066     a
#> 509        0 231070     b
#> 510        0 231083     b
#> 511        0 231095     b
#> 512        0 231100     b
#> 513        0 240082     a
#> 514        0 240092     a
#> 515        0 240097     a
#> 516        0 240297     a
#> 517        0 240328     a
#> 518        0 240502     b
#> 519        0 240675     b
#> 520        0 240752     a
#> 521        0 240801     a
#> 522        0 240875     a
#> 523        0 240880     b
#> 524        0 241150     a
#> 525        0 241191     a
#> 526        0 241210     a
#> 527        0 241243     b
#> 528        0 241284     a
#> 529        0 241347     a
#> 530        0 250023     a
#> 531        0 250087     b
#> 532        0 250091     b
#> 533        0 250106     b
#> 534        0 250126     b
#> 535        0 250127     b
#> 536        0 250143     a
#> 537        0 250194     a
#> 538        0 250203     b
#> 539        0 250224     b
#> 540        0 250225     a
#> 541        0 250231     b
#> 542        0 250235     b
#> 543        0 250240     a
#> 544        0 250288     a
#> 545        0 250291     a
#> 546        0 250299     b
#> 547        0 250300     a
#> 548        0 250301     b
#> 549        0 250304     b
#> 550        0 250406     b
#> 551        0 250419     a
#> 552        0 250422     b
#> 553        0 250426     a
#> 554        0 250433     b
#> 555        0 250436     b
#> 556        0 250443     b
#> 557        0 250447     a
#> 558        0 250454     b
#> 559        0 250457     b
#> 560        0 250459     a
#> 561        0 250461     b
#> 562        0 250464     b
#> 563        0 250465     a
#> 564        0 250471     a
#> 565        0 250472     a
#> 566        0 250476     a
#> 567        0 250481     b
#> 568        0 250482     a
#> 569        0 250485     b
#> 570        0 250490     b
#> 571        0 250516     a
#> 572        0 251021     a
#> 573        0 251023     a
#> 574        0 251026     b
#> 575        0 251038     a
#> 576        0 251048     a
#> 577        0 251058     a
#> 578        0 251059     b
#> 579        0 251060     b
#> 580        0 251068     a
#> 581        0 260157     a
#> 582        0 260722     a
#> 583        0 260735     a
#> 584        0 260757     b
#> 585        0 260817     a
#> 586        0 260828     a
#> 587        0 260867     a
#> 588        0 260879     b
#> 589        0 260897     a
#> 590        0 260920     b
#> 591        0 260932     b
#> 592        0 260934     a
#> 593        0 260991     a
#> 594        0 261047     b
#> 595        0 261050     a
#> 596        0 261052     a
#> 597        0 261053     b
#> 598        0 261065     b
#> 599        0 261078     b
#> 600        0 261085     b
#> 601        0 261107     a
#> 602        0 270031     a
#> 603        0 270042     b
#> 604        0 270056     b
#> 605        0 270060     a
#> 606        0 270082     a
#> 607        0 270137     a
#> 608        0 270737     a
#> 609        0 270743     a
#> 610        0 270744     b
#> 611        0 270750     b
#> 612        0 270751     b
#> 613        0 270752     a
#> 614        0 270753     b
#> 615        0 270755     a
#> 616        0 270758     b
#> 617        0 270761     a
#> 618        0 270826     b
#> 619        0 270827     a
#> 620        0 270828     b
#> 621        0 270830     b
#> 622        0 270832     b
#> 623        0 270833     b
#> 624        0 270834     a
#> 625        0 270837     b
#> 626        0 270840     b
#> 627        0 270845     b
#> 628        0 270846     a
#> 629        0 270847     a
#> 630        0 270851     a
#> 631        0 270853     a
#> 632        0 270858     a
#> 633        0 270859     b
#> 634        0 270862     b
#> 635        0 270873     b
#> 636        0 270877     a
#> 637        0 270885     b
#> 638        0 270888     a
#> 639        0 270889     a
#> 640        0 270892     b
#> 641        0 270893     b
#> 642        0 270983     b
#> 643        0 270984     b
#> 644        0 271001     a
#> 645        0 271365     b
#> 646        0 271390     a
#> 647        0 271735     a
#> 648        0 280226     b
#> 649        0 280236     b
#> 650        0 280272     a
#> 651        0 280286     a
#> 652        0 290062     a
#> 653        0 300113     a
#> 654        0 300233     a
#> 655        0 300500     a
#> 656        0 300507     b
#> 657        0 300522     b
#> 658        0 300903     a
#> 659        0 300907     b
#> 660        0 300938     b
#> 661        0 310038     a
#> 662        0 310171     a
#> 663        0 310182     b
#> 664        0 310211     b
#> 665        0 310215     b
#> 666        0 310216     b
#> 667        0 310225     a
#> 668        0 310228     b
#> 669        0 310234     b
#> 670        0 310236     a
#> 671        0 310237     a
#> 672        0 310238     a
#> 673        0 310244     a
#> 674        0 310713     a
#> 675        0 310739     b
#> 676        0 310761     a
#> 677        0 310763     b
#> 678        0 310767     a
#> 679        0 320088     b
#> 680        0 320286     a
#> 681        0 320287     b
#> 682        0 320288     a
#> 683        0 320291     b
#> 684        0 320296     a
#> 685        0 320297     a
#> 686        0 320298     a
#> 687        0 320318     b
#> 688        0 320319     a
#> 689        0 320325     b
#> 690        0 320326     a
#> 691        0 320357     b
#> 692        0 320362     a
#> 693        0 320365     b
#> 694        0 320369     b
#> 695        0 320392     b
#> 696        0 320393     a
#> 697        0 320396     b
#> 698        0 320412     b
#> 699        0 320413     b
#> 700        0 320418     b
#> 701        0 320420     b
#> 702        0 320431     a
#> 703        0 320447     a
#> 704        0 320450     a
#> 705        0 320464     b
#> 706        0 320472     b
#> 707        0 320473     a
#> 708        0 320475     b
#> 709        0 320502     a
#> 710        0 320531     b
#> 711        0 320534     b
#> 712        0 320536     b
#> 713        0 330074     a
#> 714        0 330156     a
#> 715        0 330234     a
#> 716        0 330240     a
#> 717        0 330241     b
#> 718        0 330244     b
#> 719        0 330247     b
#> 720        0 330248     b
#> 721        0 330255     a
#> 722        0 330257     a
#> 723        0 330259     b
#> 724        0 330264     b
#> 725        0 330265     b
#> 726        0 330267     b
#> 727        0 330274     a
#> 728        0 330282     a
#> 729        0 330283     b
#> 730        0 340034     a
#> 731        0 340131     b
#> 732        0 340298     b
#> 733        0 340602     b
#> 734        0 340606     b
#> 735        0 340684     a
#> 736        0 340718     b
#> 737        0 340731     b
#> 738        0 340735     a
#> 739        0 340744     b
#> 740        0 340767     a
#> 741        0 430074     a
#> 742        0 490308     b
#> 743        0 520014     b
#> 744        0 520015     b
#> 745        0 520022     a
#> 746        0 520031     b
#> 747        0 530009     a
#> 748        0 530014     b
#> 749        0 530016     b
#> 750        0 530027     a
#> 751        0 530031     a
#> 752        0 530032     b
#> 753        0 540012     b
#> 754        0 540014     a
#> 755        0 540015     b
#> 756        0 540021     b
#> 757        0 540022     a
#> 758        0 540024     b
#> 759        0 570012     b
#> 760        0 570013     b
#> 761        0 570015     b
#> 762        0 580009     a
#> 763        0 580011     b
#> 764        0 580013     a
#> 765        0 580016     a
#> 766        0 580018     b
#> 767        0 580020     a
#> 768        0 580021     a
#> 769        0 580027     b
#> 770        0 580028     b
#> 771        0 580030     b
#> 772        0 580035     b
#> 773        0 580040     a
#> 774        0 610015     b
#> 775        0 610017     a
#> 776        0 610018     a
#> 777        0 610023     a
#> 778        0 610025     a
#> 779        0 610028     a
#> 780        0 610034     a
#> 781        0 610035     b
#> 782        0 610038     b
#> 783        0 610042     b
#> 784        0 620020     b
#> 785        0 620022     b
#> 786        0 620023     b
#> 787        0 620025     a
#> 788        0 620027     b
#> 789        0 620038     a
#> 790        0 620042     b
#> 791        0 620045     b
#> 792        0 620046     a
#> 793        0 620054     b
#> 794        0 620058     a
#> 795        0 620420     a
#> 796        0 630014     b
#> 797        0 630016     b
#> 798        0 630021     b
#> 799        0 630022     b
#> 800        0 630023     a
#> 801        0 630026     a
#> 802        0 630031     a
#> 803        0 630034     a
#> 804        0 630036     a
#> 805        0 630038     b
#> 806        0 630042     b
#> 807        0 630043     a
#> 808        0 640009     a
#> 809        0 640012     a
#> 810        0 640018     b
#> 811        0 640020     a
#> 812        0 640021     a
#> 813        0 640026     b
#> 814        0 640027     b
#> 815        0 640030     b
#> 816        0 900017     b
#> 817        0 910031     b
#> 818        0 910038     a
#> 819        0 910039     a
#> 820        0 910044     a
#> 821        0 910045     b
#> 822        0 910056     a
#> 823        0 920031     a
#> 824        0 920043     a
#> 825        0 920044     b
#> 826        0 920046     a
#> 827        0 920051     a
#> 828        0 920052     a
#> 829        0 940519     a
#> 830        0 940522     a
#> 831        0 940528     a
#> 832        0 940529     b
#> 833        0 940531     a
#> 834        0 940532     b
#> 835        0 940535     b
#> 836        0 940536     b
#> 837        0 940540     a
#> 838        0 940547     a
#> 839        0 940549     b
#> 840        0 950027     a
#> 841        0 950034     b
#> 842        0 950036     a
#> 843        0 950037     a
#> 844        0 950051     b
#> 845        0 950056     a
#> 846        0 950062     a
#> 847        0 950071     a
#> 848        0 950075     a
#> 849        0 950078     a
#> 850        0 960011     a
#> 851        0 960016     a
#> 852        0 960020     a
#> 853        0 960021     a
#> 854        0 960022     a
#> 855        0 960028     b
#> 856        0 960031     a
#> 857        0 970026     a
#> 858        0 970027     a
#> 859        0 970462     b
#> 860        0 970466     a
#> 861        0 970475     a
#> 862        0 980011     b
#> 863        0 980022     a
#> 864        0 980023     a
#> 865        0 980027     a
#> 866        0 980042     a
#> 867        0 980046     a
#> 868        0 990026     a
#> 869        0 990071     a
#> 870        1  10368     b
#> 871        1  10476     a
#> 872        1  10899     b
#> 873        1  10908     a
#> 874        1  10913     b
#> 875        1  10918     a
#> 876        1  10920     a
#> 877        1  10929     b
#> 878        1  10934     b
#> 879        1  10947     b
#> 880        1  10961     a
#> 881        1  11127     a
#> 882        1  11200     a
#> 883        1  11445     b
#> 884        1  11450     a
#> 885        1  11671     a
#> 886        1  11954     a
#> 887        1  12001     b
#> 888        1  20508     a
#> 889        1  20706     b
#> 890        1  20724     a
#> 891        1  40422     a
#> 892        1  40758     b
#> 893        1  41171     b
#> 894        1  41190     a
#> 895        1  50207     a
#> 896        1  50609     b
#> 897        1  50610     b
#> 898        1  50618     a
#> 899        1  50632     a
#> 900        1  50644     b
#> 901        1  50650     b
#> 902        1  60151     b
#> 903        1  60307     a
#> 904        1  60625     a
#> 905        1  60627     a
#> 906        1  60884     a
#> 907        1  60891     a
#> 908        1  60894     a
#> 909        1  60926     a
#> 910        1  60927     b
#> 911        1  60928     a
#> 912        1  70477     b
#> 913        1  70491     b
#> 914        1  70541     b
#> 915        1  70787     b
#> 916        1  70788     b
#> 917        1  71339     b
#> 918        1  71349     a
#> 919        1  71354     b
#> 920        1  71360     b
#> 921        1  71367     b
#> 922        1  71373     a
#> 923        1  71375     a
#> 924        1  71377     b
#> 925        1  71387     a
#> 926        1  71391     a
#> 927        1  80282     b
#> 928        1  80413     a
#> 929        1  81450     a
#> 930        1  81457     b
#> 931        1  81532     b
#> 932        1  90683     b
#> 933        1  90689     b
#> 934        1  90692     a
#> 935        1 110307     b
#> 936        1 110373     b
#> 937        1 110652     a
#> 938        1 110746     a
#> 939        1 120921     b
#> 940        1 140196     b
#> 941        1 140228     a
#> 942        1 140237     a
#> 943        1 140260     b
#> 944        1 140287     b
#> 945        1 140732     b
#> 946        1 140748     a
#> 947        1 140768     a
#> 948        1 140777     b
#> 949        1 140800     b
#> 950        1 150062     b
#> 951        1 150093     a
#> 952        1 150108     a
#> 953        1 150160     a
#> 954        1 150168     a
#> 955        1 150431     b
#> 956        1 150441     b
#> 957        1 150463     a
#> 958        1 150468     a
#> 959        1 150472     a
#> 960        1 170342     b
#> 961        1 170472     a
#> 962        1 170541     a
#> 963        1 170690     a
#> 964        1 180692     a
#> 965        1 180698     b
#> 966        1 181961     a
#> 967        1 181963     b
#> 968        1 181964     a
#> 969        1 181990     b
#> 970        1 190196     b
#> 971        1 210612     a
#> 972        1 210645     b
#> 973        1 210708     a
#> 974        1 211070     a
#> 975        1 211328     b
#> 976        1 220036     a
#> 977        1 230024     b
#> 978        1 230064     b
#> 979        1 230080     b
#> 980        1 230087     b
#> 981        1 230103     a
#> 982        1 230278     a
#> 983        1 230286     a
#> 984        1 230292     b
#> 985        1 230411     a
#> 986        1 230537     b
#> 987        1 230578     a
#> 988        1 230686     a
#> 989        1 231023     a
#> 990        1 231046     b
#> 991        1 231065     a
#> 992        1 240019     a
#> 993        1 240797     a
#> 994        1 240892     b
#> 995        1 240934     b
#> 996        1 241099     b
#> 997        1 241154     b
#> 998        1 241339     b
#> 999        1 241359     b
#> 1000       1 241370     a
#> 1001       1 250144     b
#> 1002       1 250163     b
#> 1003       1 250404     a
#> 1004       1 251029     a
#> 1005       1 251031     a
#> 1006       1 251033     b
#> 1007       1 251041     b
#> 1008       1 251049     b
#> 1009       1 251051     a
#> 1010       1 260244     b
#> 1011       1 260693     a
#> 1012       1 270074     a
#> 1013       1 270129     b
#> 1014       1 270738     b
#> 1015       1 270747     b
#> 1016       1 270748     b
#> 1017       1 270854     b
#> 1018       1 270863     a
#> 1019       1 270865     b
#> 1020       1 270934     a
#> 1021       1 300010     a
#> 1022       1 300027     b
#> 1023       1 300930     b
#> 1024       1 310233     a
#> 1025       1 310769     a
#> 1026       1 320348     a
#> 1027       1 320351     b
#> 1028       1 320408     b
#> 1029       1 320409     a
#> 1030       1 320410     a
#> 1031       1 320411     a
#> 1032       1 320415     b
#> 1033       1 320419     a
#> 1034       1 330164     b
#> 1035       1 330236     a
#> 1036       1 330256     a
#> 1037       1 330275     a
#> 1038       1 340708     b
#> 1039       1 340713     b
#> 1040       1 340748     b
#> 1041       1 520019     a
#> 1042       1 570009     b
#> 1043       1 620010     b
#> 1044       1 620018     b
#> 1045       1 620037     a
#> 1046       1 620416     b
#> 1047       1 630012     a
#> 1048       1 630030     b
#> 1049       1 900031     a
#> 1050       1 910028     a
#> 1051       1 960014     b
#> 1052       1 960026     a
#> 1053       1 970032     b
#> 1054       1 980009     b